+++ src/main/java/id/pkl/raport/repository/UsersRepository.java
@@ -1,5 +1,7 @@
 package id.pkl.raport.repository;
 
+import java.io.Serializable;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 
 import id.pkl.raport.entity.Users;
 
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 	@Query("SELECT users FROM Users users WHERE LOWER(users.username) LIKE LOWER(CONCAT('%', :search, '%'))"
 			+ "OR LOWER(users.password) LIKE LOWER(CONCAT('%', :search, '%'))")
 	Page<Users> findBySearch(@Param("search") String searchField, Pageable pageable);
